Users can not delete Threads.

+++ app/Http/Controllers/ThreadsController.php
@@ -109,6 +109,10 @@ class ThreadsController extends Controller
      */
     public function destroy(Channel $channel, Thread $thread)
     {
+        $this->authorize("update", $thread);
+        if($thread->user_id != auth()->id()) {
+            abort(403, "You do not have permission to do this.");
+        }
         $thread->delete();
 
         if(request()->wantsJson()) {
